Task: Added the StudentFactory domaim and tests implementation
